updated inputmanager, renderer, and collision manager to support multiple players in the game world, added world border

+++ Src/TileState.h
@@ -6,7 +6,7 @@
 struct TileState
 {
     enum TileType TileType;
-    Uint8 Health;
+    int Health;
 };
 
 #endif
